Move actions related to viewing details and the loading bar into FlowActions
