+++ client/src/result/granules/list/GranuleListContainer.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux'
-import { toggleGranuleFocus } from '../../../actions/SearchRequestActions'
+import { toggleGranuleFocus } from '../../../actions/FlowActions'
 import GranuleList from './GranuleListComponent'
 
 const mapStateToProps = (state) => {
