+++ client/test/reducers/ui/granuleDetailsSpec.js
@@ -1,7 +1,6 @@
 import '../../specHelper'
-import Immutable from 'seamless-immutable'
 import { granuleDetails, initialState } from '../../../src/reducers/ui/granuleDetails'
-import { toggleGranuleFocus } from '../../../src/actions/SearchRequestActions'
+import { toggleGranuleFocus } from '../../../src/actions/FlowActions'
 
 describe('The granuleDetails reducer', function() {
   it('has a default state', function () {
